GridクラスのセルをジェネリクスにすることでセルクラスをTopologyCalculator.csに移動した
